email-agent/pyproject.toml
[project]
name = "email-agent"
version = "0.1.0"
description = "A Gmail agent that reads, categorizes, and acts on your inbox using Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
